Using suite enum for suite list
