+++ test/ts/src/unary.spec.ts
@@ -14,9 +14,9 @@ import {
 import {FailService, TestService} from "../_proto/improbable/grpcweb/test/test_pb_service";
 import {DEBUG, UncaughtExceptionListener} from "./util";
 import {
-  conditionallyRunTestSuite, headerTrailerCombos, runWithHttp1AndHttp2,
-  runWithSupportedTransports, SuiteEnum
+  headerTrailerCombos, runWithHttp1AndHttp2, runWithSupportedTransports
 } from "./testRpcCombinations";
+import { conditionallyRunTestSuite, SuiteEnum } from "../suiteUtils";
 
 conditionallyRunTestSuite(SuiteEnum.unary, () => {
   runWithHttp1AndHttp2(({ testHostUrl, corsHostUrl, unavailableHost, emptyHost}) => {
